+++ yabu_cmd/__init__.py
@@ -1,3 +1,2 @@
 from .controller import *
-from .model import *
 from .exceptions import *
