--- yabu_cmd/model/__init__.py
@@ -1,2 +1,0 @@
-from .backup_manager import *
-from .preset_manager import *
